// app/client/src/api/ApiConstants.ts
export const REQUEST_TIMEOUT_MS = 20000;

export enum API_STATUS_CODES {
  REQUEST_NOT_AUTHORISED = 401,
  RESOURCE_NOT_FOUND = 404,
  SERVER_ERROR = 502,
  SERVER_UNAVAILABLE = 503,
  REQUEST_FORBIDDEN = 403,
}

export const SERVER_ERROR_CODES = {
  RESOURCE_NOT_FOUND: ['AE-SCL-4004', 'AE-BAD-4000', 'AE-APP-4028', 'AE-APP-4013'],
};

export enum ERROR_CODES {
  PAGE_NOT_FOUND = 'PAGE_NOT_FOUND',
  SERVER_ERROR = 'SERVER_ERROR',
  REQUEST_NOT_AUTHORISED = 'REQUEST_NOT_AUTHORIZED',
  REQUEST_TIMEOUT = 'REQUEST_TIMEOUT',
  FAILED_TO_CORRECT_BINDING = 'FAILED_TO_CORRECT_BINDING',
  REQUEST_FORBIDDEN = 'REQUEST_FORBIDDEN',
  CYPRESS_DEBUG = 'CYPRESS_DEBUG',
}